requirements/requirements.txt
alembic==0.8.4
BeautifulSoup==3.2.1
blinker==1.4
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
Pillow==3.1.1
python-editor==0.5
qiniu==7.0.6
requests==2.9.1
SQLAlchemy==1.0.12
Werkzeug==0.11.4
wheel==0.24.0
WTForms==2.1
MySQL-python==1.2.5
